直通先の路線を乗り換え扱いにしない
